add SelectPickerUseCase
