+++ example-app/SantokuApp/src/components/picker/SelectPicker.ios.tsx
@@ -1,106 +1,49 @@
-import React, {useCallback, useMemo, useState} from 'react';
+import React from 'react';
 import {Pressable, View} from 'react-native';
 import {Input} from 'react-native-elements';
 
 import {PickerAccessory} from './PickerAccessory';
-import {
-  DEFAULT_ENTERING as DEFAULT_PICKER_BACKDROP_ENTERING,
-  DEFAULT_EXITING as DEFAULT_PICKER_BACKDROP_EXITING,
-  PickerBackdrop,
-} from './PickerBackdrop';
-import {
-  DEFAULT_ENTERING as DEFAULT_PICKER_CONTAINER_ENTERING,
-  DEFAULT_EXITING as DEFAULT_PICKER_CONTAINER_EXITING,
-  PickerContainer,
-} from './PickerContainer';
-import {Item, ItemWithKey, SelectPickerProps} from './SelectPicker';
+import {PickerBackdrop} from './PickerBackdrop';
+import {PickerContainer} from './PickerContainer';
+import {SelectPickerProps} from './SelectPicker';
 import {SelectPickerItems} from './SelectPickerItems';
+import {useSelectPickerUseCase} from './useSelectPickerUseCase';
 
-const DEFAULT_PLACEHOLDER_COLOR = 'grey';
-export const DEFAULT_DURATION = 500;
+export type SelectPickerPropsIOS<ItemT> = Omit<SelectPickerProps<ItemT>, 'useNativeAndroidPickerStyle'>;
 
-const handlePlaceholder = <ItemT extends unknown>(placeholder?: Item<ItemT>): (Item<ItemT> | ItemWithKey<ItemT>)[] => {
-  return placeholder ? [{color: DEFAULT_PLACEHOLDER_COLOR, ...placeholder}] : [];
-};
-
-type SelectPickerIOS<ItemT> = Omit<SelectPickerProps<ItemT>, 'useNativeAndroidPickerStyle'>;
-
-export const SelectPicker = <ItemT extends unknown>({
-  items,
-  selectedItemKey,
-  onSelectedItemChange,
-  onDelete,
-  onCancel,
-  onDone,
-  placeholder,
-  textInputProps,
-  keyExtractor,
-  headerComponent,
-  itemsComponent,
-  pickerItemsContainerProps,
-  pickerProps,
-  backdropProps: {entering: backdropEntering, exiting: backdropExiting, onPress, ...backdropProps} = {},
-  containerProps: {entering: containerEntering, exiting: containerExiting, ...containerProps} = {},
-  pickerAccessoryProps,
-}: SelectPickerIOS<ItemT>) => {
-  const [isVisible, setIsVisible] = useState(false);
-  const close = useCallback(() => setIsVisible(false), []);
-  const handleBackdropPress = useCallback(() => {
-    onPress?.();
-    close();
-  }, [close, onPress]);
-  const pickerBackdropEntering = useMemo(
-    () => backdropEntering ?? DEFAULT_PICKER_BACKDROP_ENTERING.duration(DEFAULT_DURATION),
-    [backdropEntering],
-  );
-  const pickerBackdropExiting = useMemo(
-    () => backdropExiting ?? DEFAULT_PICKER_BACKDROP_EXITING.duration(DEFAULT_DURATION),
-    [backdropExiting],
-  );
-  const pickerContainerEntering = useMemo(
-    () => containerEntering ?? DEFAULT_PICKER_CONTAINER_ENTERING.duration(DEFAULT_DURATION),
-    [containerEntering],
-  );
-  const pickerContainerExiting = useMemo(
-    () => containerExiting ?? DEFAULT_PICKER_CONTAINER_EXITING.duration(DEFAULT_DURATION),
-    [containerExiting],
-  );
-  const itemsWithPlaceholder = useMemo(() => handlePlaceholder(placeholder).concat(items), [items, placeholder]);
-  const getSelectedItem = useCallback(
-    (key?: React.Key) => {
-      if (key) {
-        const found = itemsWithPlaceholder.find(item => item.key === key);
-        return found ?? itemsWithPlaceholder.find(item => item.value === key);
-      }
-      if (placeholder) {
-        return placeholder;
-      }
-      return undefined;
-    },
-    [itemsWithPlaceholder, placeholder],
-  );
-  const onValueChange = useCallback(
-    (key: React.Key, index: number) => {
-      const selectedItem = getSelectedItem(key);
-      onSelectedItemChange?.(index, selectedItem?.value, key);
-    },
-    [getSelectedItem, onSelectedItemChange],
-  );
-
-  const selectedItem = useMemo(() => getSelectedItem(selectedItemKey), [getSelectedItem, selectedItemKey]);
-
-  const handleDelete = useCallback(() => {
-    onDelete?.(selectedItem);
-    close();
-  }, [close, onDelete, selectedItem]);
-  const handleCancel = useCallback(() => {
-    onCancel?.(selectedItem);
-    close();
-  }, [close, onCancel, selectedItem]);
-  const handleDone = useCallback(() => {
-    onDone?.(selectedItem);
-    close();
-  }, [close, onDone, selectedItem]);
+export const SelectPicker = <ItemT extends unknown>(props: SelectPickerPropsIOS<ItemT>) => {
+  const {
+    isVisible,
+    itemsWithPlaceholder,
+    selectedItem,
+    handleBackdropPress,
+    pickerBackdropEntering,
+    pickerBackdropExiting,
+    pickerContainerEntering,
+    pickerContainerExiting,
+    onValueChange,
+    open,
+    handleDelete,
+    handleCancel,
+    handleDone,
+  } = useSelectPickerUseCase<ItemT>(props);
+  const {
+    selectedItemKey,
+    textInputProps,
+    keyExtractor,
+    headerComponent,
+    itemsComponent,
+    pickerItemsContainerProps,
+    pickerProps,
+    backdropProps: {
+      entering: backdropEntering,
+      exiting: backdropExiting,
+      onPress: onBackdropPress,
+      ...backdropProps
+    } = {},
+    containerProps: {entering: containerEntering, exiting: containerExiting, ...containerProps} = {},
+    pickerAccessoryProps,
+  } = props;
 
   return (
     <>
@@ -139,7 +82,7 @@ export const SelectPicker = <ItemT extends unknown>({
           )}
         </PickerContainer>
       </PickerBackdrop>
-      <Pressable onPress={() => setIsVisible(true)}>
+      <Pressable onPress={open}>
         <View pointerEvents="box-only">
           <Input value={selectedItem?.inputLabel ?? selectedItem?.label} editable={false} {...textInputProps} />
         </View>
